+++ sdk/resourcemanager/network/armnetwork/models_serde.go
@@ -7456,6 +7456,33 @@ func (b *BastionShareableLinkListResult) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type BastionShareableLinkTokenListRequest.
+func (b BastionShareableLinkTokenListRequest) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "tokens", b.Tokens)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type BastionShareableLinkTokenListRequest.
+func (b *BastionShareableLinkTokenListRequest) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", b, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "tokens":
+			err = unpopulate(val, "Tokens", &b.Tokens)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", b, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type BgpConnection.
 func (b BgpConnection) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)